sql: fix displaying negative second zone offsets
timestamptz->text casts displayed sub-minute zone offsets west of UTC
incorrectly. The America/Chicago LMT offset as described in tzdata
is -21036 seconds (-05:50:36); `-21036 % 60` is `-36`. A correct
implementation displays the absolute value of this remainder instead of
adding a negative remainder to 60, and can be found in the Go standard
library instead of being reimplemented here.

Fixes: 07219f63fa19 ("sql: fix displaying second zone offsets")