work around illumos#16834 in test suite
/usr/share/lib/zoneinfo on illumos is generated by a timezone compiler
that is unaware of how to write 64-bit entries; thus America/Chicago is
missing the entry that puts times prior to November 18, 1883, in local
mean time, which is required for the sub-minute timezone offset tests to
pass as-written.

To continue testing this behavior we instead use America/Halifax,
which (according to tzdata) switched from LMT to standard time in 1902,
shortly after the world unknowingly entered the period of time that can
be represented on a computer operating system that would be invented
around 70 years later.