remove most calls to deprecated rand.Seed
Since Go 1.22, the default unseeded global RNG is more secure:
https://go.dev/blog/chacha8rand

However when the RNG is seeded it falls back to the old implementation.

One call to rand.Seed remains in randutil.SeedForTests. This is part
of the COCKROACH_RANDOM_SEED infrastructure for recording what seed was
used for a particular test failure so it could be reproduced. (Had the
global RNG not been seedable in the first place, perhaps there would be
more sensible infrastructure in place here.) We will leave this in place
until Go 1.24 when rand.Seed becomes a no-op by default.